+++ cookiemonster/processor/basic_processing.py
@@ -25,17 +25,19 @@ import logging
 import time
 import traceback
 from concurrent.futures import ThreadPoolExecutor
+from datetime import datetime
 from typing import Sequence
 
-from hgicommon.data_source import DataSource
-
-from cookiemonster.common.models import Cookie
+from cookiemonster.common.models import Cookie, Enrichment
 from cookiemonster.cookiejar import CookieJar
 from cookiemonster.logging.logger import PythonLoggingLogger, Logger
 from cookiemonster.processor._enrichment import EnrichmentManager
 from cookiemonster.processor._rules import RuleQueue
-from cookiemonster.processor.models import Rule, EnrichmentLoader
-from cookiemonster.processor.processing import ProcessorManager, Processor
+from cookiemonster.processor.json_convert import RuleApplicationLogJSONEncoder
+from cookiemonster.processor.models import Rule, EnrichmentLoader, RuleApplicationLog
+from cookiemonster.processor.processing import ProcessorManager, Processor, RULE_APPLICATION
+from hgicommon.collections import Metadata
+from hgicommon.data_source import DataSource
 
 _MEASUREMENT_PROCESSING_COUNT = "processing"
 _MEASUREMENT_GET_NEXT_COUNT = "get_next_for_processing"
@@ -46,6 +48,8 @@ class BasicProcessor(Processor):
     """
     Simple processor for a single Cookie.
     """
+    _RULE_APPLICATION_LOG_JSON_ENCODER = RuleApplicationLogJSONEncoder()
+
     def __init__(self, cookie_jar: CookieJar, rules: Sequence[Rule], enrichment_loaders: Sequence[EnrichmentLoader]):
         """
         Constructor.
@@ -65,10 +69,20 @@ class BasicProcessor(Processor):
             rule = rule_queue.get_next()
             isolated_cookie = copy.deepcopy(cookie)
             try:
-                if rule.matches(isolated_cookie):
+                matches = rule.matches(isolated_cookie)
+                if matches:
                     terminate = rule.execute_action(isolated_cookie)
-            except Exception:
+            except:
+                matches = False
                 logging.error("Error applying rule; Rule: %s; Error: %s" % (rule, traceback.format_exc()))
+
+            if matches:
+                log = RuleApplicationLog(rule.id, terminate)
+                log_as_dict = BasicProcessor._RULE_APPLICATION_LOG_JSON_ENCODER.default(log)
+                enrichment = Enrichment(RULE_APPLICATION, datetime.now(), Metadata(log_as_dict))
+                self.cookie_jar.enrich_cookie(cookie.identifier, enrichment, mark_for_processing=False)
+                # Update in-memory copy of cookie
+                cookie.enrichments.append(enrichment)
             rule_queue.mark_as_applied(rule)
 
         return terminate
@@ -80,9 +94,7 @@ class BasicProcessor(Processor):
         enrichment = enrichment_manager.next_enrichment(cookie)
 
         if enrichment is None:
-            logging.info("Cannot enrich cookie with identifier \"%s\" any further - notifying listeners"
-                         % cookie.identifier)
-            # TODO: Should anything else be done here?
+            logging.info("Cannot enrich cookie with identifier \"%s\" any further" % cookie.identifier)
         else:
             logging.info("Applying enrichment from source \"%s\" to cookie with identifier \"%s\""
                          % (enrichment.source, cookie.identifier))