+++ cookiemonster/tests/processor/_rules/name_match_rule.py
@@ -44,5 +44,5 @@ def _action(cookie: Cookie, context: Context) -> bool:
 
 _priority = Priority.MAX_PRIORITY
 
-_rule = Rule(MagicMock(side_effect=_matches), MagicMock(side_effect=_action), _priority, NAME_MATCH_RULE_ID)
+_rule = Rule(MagicMock(side_effect=_matches), MagicMock(side_effect=_action), NAME_MATCH_RULE_ID, _priority)
 register(_rule)