+++ cookiemonster/tests/processor/_rules/no_match_rule.py
@@ -41,5 +41,5 @@ def _action(cookie: Cookie, context: Context) -> bool:
 
 _priority = Priority.MAX_PRIORITY
 
-_rule = Rule(MagicMock(side_effect=_matches), MagicMock(side_effect=_action), _priority, NO_MATCH_RULE_ID)
+_rule = Rule(MagicMock(side_effect=_matches), MagicMock(side_effect=_action), NO_MATCH_RULE_ID, _priority)
 register(_rule)