+++ cookiemonster/tests/processor/test_basic_processing.py
@@ -25,18 +25,21 @@ from datetime import datetime
 from threading import Semaphore, Thread, Lock
 from unittest.mock import MagicMock, call
 
+from cookiemonster.common.context import Context
+from cookiemonster.common.models import Cookie, Notification, Enrichment
+from cookiemonster.cookiejar.in_memory_cookiejar import InMemoryCookieJar
+from cookiemonster.processor.basic_processing import BasicProcessor, BasicProcessorManager
+from cookiemonster.processor.json_convert import RuleApplicationLogJSONDecoder
+from cookiemonster.processor.models import Rule, EnrichmentLoader
+from cookiemonster.processor.models import RuleApplicationLog
+from cookiemonster.processor.processing import RULE_APPLICATION
+from cookiemonster.tests.processor._mocks import create_magic_mock_cookie_jar
 from hgicommon.collections import Metadata
 from hgicommon.data_source import ListDataSource
 from hgicommon.mixable import Priority
 
-from cookiemonster.common.models import Cookie, Notification, Enrichment
-from cookiemonster.common.context import Context
-from cookiemonster.cookiejar.in_memory_cookiejar import InMemoryCookieJar
-from cookiemonster.processor.basic_processing import BasicProcessor, BasicProcessorManager
-from cookiemonster.processor.models import Rule, EnrichmentLoader
-from cookiemonster.tests.processor._mocks import create_magic_mock_cookie_jar
-
 COOKIE_IDENTIFIER = "/my/cookie"
+RULE_IDENTIFIER = "my_rule"
 SAMPLE_ENRICHMENT = Enrichment("sample", datetime(year=2000, month=1, day=1), Metadata())
 
 
@@ -46,8 +49,9 @@ class TestBasicProcessor(unittest.TestCase):
     """
     def setUp(self):
         self.cookie_jar = InMemoryCookieJar()
-        self.rules = [Rule(lambda *args: False, MagicMock()) for _ in range(10)]
-        self.cookie = Cookie(COOKIE_IDENTIFIER)
+        self.rules = [Rule(lambda *args: False, MagicMock(), RULE_IDENTIFIER) for _ in range(10)]
+        self.cookie_jar.enrich_cookie(COOKIE_IDENTIFIER, Metadata())
+        self.cookie = self.cookie_jar.fetch_cookie(COOKIE_IDENTIFIER)
         self.processor = BasicProcessor(self.cookie_jar, [], [])
 
     def test_evaluate_rules_with_cookie_when_no_rules(self):
@@ -65,14 +69,17 @@ class TestBasicProcessor(unittest.TestCase):
 
     def test_evaluate_rules_with_cookie_when_matched_rules_and_no_termination(self):
         extra_rules = [
-            Rule(lambda *args: True, MagicMock(return_value=False)),
-            Rule(lambda *args: True, MagicMock(return_value=False))
+            Rule(lambda *args: True, MagicMock(return_value=False), RULE_IDENTIFIER),
+            Rule(lambda *args: True, MagicMock(return_value=False), RULE_IDENTIFIER)
         ]
         self.processor.rules = self.rules + extra_rules
         halt = self.processor.evaluate_rules_with_cookie(self.cookie)
 
         for rule in self.rules:
             rule._action.assert_not_called()
+        for rule in extra_rules:
+            self.assertEqual(rule._action.call_count, 1)
+            self.assertEqual(rule._action.call_args[0][0].identifier, self.cookie.identifier)
         self.assertFalse(halt)
 
     def test_evaluate_rules_with_cookie_when_matched_rules_and_termination(self):
@@ -108,16 +115,39 @@ class TestBasicProcessor(unittest.TestCase):
                 change_detected_in_next_rule = True
 
         self.processor.rules = [
-            Rule(cookie_changer, self.rules[0]._action, priority=Priority.MAX_PRIORITY),
-            Rule(record_fail_if_changed, self.rules[0]._action, priority=Priority.MIN_PRIORITY)
+            Rule(cookie_changer, self.rules[0]._action, RULE_IDENTIFIER, priority=Priority.MAX_PRIORITY),
+            Rule(record_fail_if_changed, self.rules[0]._action, RULE_IDENTIFIER, priority=Priority.MIN_PRIORITY)
         ]
         self.processor.process_cookie(self.cookie)
 
         self.assertFalse(change_detected_in_next_rule)
 
+    def test_evaluate_rules_with_cookie_when_no_matched__does_not_enrich_with_rule_application_log(self):
+        self.processor.rules = self.rules
+        _ = self.processor.evaluate_rules_with_cookie(self.cookie)
+
+        cookie = self.cookie_jar.fetch_cookie(self.cookie.identifier)
+        self.assertEqual(len(cookie.enrichments), 1)
+
+    def test_evaluate_rules_with_cookie_when_matched_enriches_with_rule_application_log(self):
+        matched_rule_identifier = "matched_rule"
+        terminates = False
+        self.processor.rules = self.rules \
+                               + [Rule(lambda *args: True, MagicMock(return_value=terminates), matched_rule_identifier)]
+        _ = self.processor.evaluate_rules_with_cookie(self.cookie)
+
+        cookie = self.cookie_jar.fetch_cookie(self.cookie.identifier)
+        self.assertGreater(len(cookie.enrichments), 1)
+        last_enrichment = cookie.enrichments[-1]
+        self.assertEqual(last_enrichment.source, RULE_APPLICATION)
+        rule_application_log = RuleApplicationLogJSONDecoder().decode_parsed(last_enrichment.metadata)  # type: RuleApplicationLog
+        self.assertEqual(rule_application_log.rule_id, matched_rule_identifier)
+        self.assertEqual(rule_application_log.terminated_processing, terminates)
+
     def test_handle_cookie_enrichment_when_matching_enrichments(self):
         self.processor.notification_receivers = [MagicMock()]
-        self.processor.enrichment_loaders = [EnrichmentLoader(lambda *args: True, lambda *args: SAMPLE_ENRICHMENT)]
+        self.processor.enrichment_loaders = [EnrichmentLoader(
+            lambda *args: True, lambda *args: SAMPLE_ENRICHMENT, RULE_IDENTIFIER)]
 
         self.processor.handle_cookie_enrichment(self.cookie)
 
@@ -170,7 +200,7 @@ class TestBasicProcessorManager(unittest.TestCase):
 
         self.cookie_jar.mark_as_complete = MagicMock(side_effect=on_complete)
 
-        self.rules.append(Rule(lambda *args: True, lambda *args: True))
+        self.rules.append(Rule(lambda *args: True, lambda *args: True, RULE_IDENTIFIER))
 
         number_to_process = 100
         for i in range(number_to_process):
@@ -202,7 +232,7 @@ class TestBasicProcessorManager(unittest.TestCase):
             rule_lock.acquire()
             return True
 
-        self.rules.append(Rule(matching_criteria, lambda *args: True))
+        self.rules.append(Rule(matching_criteria, lambda *args: True, RULE_IDENTIFIER))
 
         self.cookie_jar.mark_for_processing(self.cookie.identifier)
         processor_manager.process_any_cookies()
@@ -215,8 +245,8 @@ class TestBasicProcessorManager(unittest.TestCase):
 
         # Change the rules for the next cookie to be processed
         self.rules.pop()
-        rule_execute_monitor = MagicMock()
-        self.rules.append(Rule(lambda *args: True, rule_execute_monitor))
+        rule_execute_monitor = MagicMock(return_value=False)
+        self.rules.append(Rule(lambda *args: True, rule_execute_monitor, RULE_IDENTIFIER))
 
         # Free the processor to complete the first cookie
         rule_lock.release()